Fail fast when tag is not defined
