+++ Tests/Unit/NetworkServer/MetricRegistryTests.cs
@@ -1,11 +1,15 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+#nullable enable
+
 namespace LoRaWan.Tests.Unit.NetworkServer
 {
     using System;
+    using System.Collections.Generic;
     using System.Diagnostics.Metrics;
     using LoRaWan.NetworkServer;
+    using Moq;
     using Xunit;
 
     public sealed class MetricRegistryTests : IDisposable
@@ -63,6 +67,81 @@ namespace LoRaWan.Tests.Unit.NetworkServer
 
             // act + assert
             Assert.Throws<ArgumentException>(() => this.meter.CreateHistogram<int>(customMetric));
+        }
+
+        [Fact]
+        public void GetTagsInOrder_Returns_Tags_When_Invoked_With_Ordered_Tags()
+        {
+            // arrange
+            var tags = new[] { "foo", "bar" };
+            var tagValues = new[] { KeyValuePair.Create("foo", (object?)"foovalue"), KeyValuePair.Create("bar", (object?)"barvalue") };
+
+            // act
+            var result = MetricExporterHelper.GetTagsInOrder(tags, tagValues);
+
+            // assert
+            Assert.Equal(new[] { "foovalue", "barvalue" }, result);
+        }
+
+        [Fact]
+        public void GetTagsInOrder_Returns_Tags_When_Invoked_With_Unordered_Tags()
+        {
+            // arrange
+            var tags = new[] { "foo", "bar" };
+            var tagValues = new[] { KeyValuePair.Create("bar", (object?)"barvalue"), KeyValuePair.Create("foo", (object?)"foovalue") };
+
+            // act
+            var result = MetricExporterHelper.GetTagsInOrder(tags, tagValues);
+
+            // assert
+            Assert.Equal(new[] { "foovalue", "barvalue" }, result);
+        }
+
+        [Fact]
+        public void GetTagsInOrder_Throws_When_Tag_Not_Found()
+        {
+            var result = Assert.Throws<LoRaProcessingException>(() => MetricExporterHelper.GetTagsInOrder(new[] { "foo" }, Array.Empty<KeyValuePair<string, object?>>()));
+            Assert.Equal(LoRaProcessingErrorCode.TagNotSet, result.ErrorCode);
+        }
+
+        [Fact]
+        public void GetTagsInOrder_Throws_When_Tag_Is_Empty()
+        {
+            const string tagName = "foo";
+            var result = Assert.Throws<LoRaProcessingException>(() => MetricExporterHelper.GetTagsInOrder(new[] { tagName }, new[] { KeyValuePair.Create(tagName, (object?)string.Empty) }));
+            Assert.Equal(LoRaProcessingErrorCode.TagNotSet, result.ErrorCode);
+        }
+
+        [Fact]
+        public void CompositeMetricExporter_Works_If_One_Exporter_Null()
+        {
+            // arrange
+            var metricExporter = new Mock<IMetricExporter>();
+
+            // act
+            using var first = new CompositeMetricExporter(null, metricExporter.Object);
+            using var second = new CompositeMetricExporter(metricExporter.Object, null);
+            first.Start();
+            second.Start();
+
+            // assert
+            metricExporter.Verify(me => me.Start(), Times.Exactly(2));
+        }
+
+        [Fact]
+        public void CompositeMetricExporter_Works_If_Both_Exporters_Defined()
+        {
+            // arrange
+            var firstExporter = new Mock<IMetricExporter>();
+            var secondExporter = new Mock<IMetricExporter>();
+
+            // act
+            using var result = new CompositeMetricExporter(firstExporter.Object, secondExporter.Object);
+            result.Start();
+
+            // assert
+            firstExporter.Verify(me => me.Start(), Times.Once);
+            secondExporter.Verify(me => me.Start(), Times.Once);
         }
 
         public void Dispose()
